demo/face-analyzer-with-llm/pyproject.toml
[project]
name = "face-analyzer-with-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gradio>=6.8.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
